refactor: LottoPrizesRecord를 클래스에서 record로 변경
