+++ src/main/java/lotto/dto/LottoPrizesRecord.java
@@ -1,22 +1,20 @@
-package lotto.domain;
+package lotto.dto;
 
 import static lotto.domain.LottoPrize.NO_PRIZE;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.Map;
+import lotto.domain.LottoPrice;
+import lotto.domain.LottoPrize;
 
-public class LottoPrizesRecord {
-    private final Map<LottoPrize, Integer> lottoPrizesMap;
-
-    public LottoPrizesRecord(Map<LottoPrize, Integer> lottoPrizesMap) {
-        this.lottoPrizesMap = lottoPrizesMap;
-    }
+public record LottoPrizesRecord(Map<LottoPrize, Integer> lottoPrizesMap) {
 
     public BigDecimal totalEarningAmount() {
         return lottoPrizesMap.entrySet().stream()
                 .filter(entry -> entry.getKey() != NO_PRIZE)
-                .map(entry -> BigDecimal.valueOf(entry.getValue()).multiply(BigDecimal.valueOf(entry.getKey().getPrizeMoney())))
+                .map(entry -> BigDecimal.valueOf(entry.getValue())
+                        .multiply(BigDecimal.valueOf(entry.getKey().getPrizeMoney())))
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 
@@ -27,9 +25,5 @@ public class LottoPrizesRecord {
                 .multiply(BigDecimal.valueOf(100))
                 .divide(lottoPrice.getPrice(), 1, RoundingMode.HALF_UP);
         return rateOfReturn;
-    }
-
-    public Map<LottoPrize, Integer> getLottoPrizesMap() {
-        return lottoPrizesMap;
     }
 }
