+++ src/test/java/lotto/domain/LottoPrizesRecordTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.math.BigDecimal;
 import java.util.Map;
 import java.util.stream.Stream;
+import lotto.dto.LottoPrizesRecord;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
